Extracted Record into a separate module
